+++ The-Adventures-Of-The-Willow-2.0/Assets/TAOTW2.0/Scripts/LevelEditor/LoadPlayLevel.cs
@@ -1283,4 +1283,13 @@ public class LoadPlayLevel : MonoBehaviour
             }
         }
     }
+
+
+    public void StopTestGameButton()
+    {
+        if (LevelEditorController.instance != null)
+        {
+            LevelEditorController.instance.TestGamePlay();
+        }
+    }
 }