+++ The-Adventures-Of-The-Willow-2.0/Assets/TAOTW2.0/Scripts/Player/PlayerMovement/UserInput.cs
@@ -29,4 +29,13 @@ public class UserInput : MonoBehaviour
     {
         playerMoveAndExtraActions.Disable();
     }
+    public void DisableInput()
+    {
+        playerMoveAndExtraActions.Disable();
+    }
+
+    public void EnableInput()
+    {
+        playerMoveAndExtraActions.Enable();
+    }
 }